requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.0
greenlet==3.1.1
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
lxml==5.3.2
numpy==2.2.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
playwright==1.50.0
pluggy==1.5.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyee==12.1.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pyparsing==3.2.1
pytest==8.3.5
pytest-base-url==2.1.0
pytest-dotenv==0.5.2
pytest-playwright==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==78.1.0
six==1.17.0
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
